+++ CHANGELOG.md
@@ -1,31 +1,50 @@
 # CHANGELOG
 
-A list of changes in application
+A list of changes in application, based in MasterVersion.MinimumVersion.PathVersion.
+Master: Larged changes (UX or application base)
+Minimum: Some changes, but nothing notable
+Patch: To fix errors
 
 ## Releases
 
+- 2.0.0
+    ADD: Material UI Components
+    FIX: All UI errors by generic components
+    CHANGED: Input forms and Theme by Material UI, New Auth Logic
+    REMOVED: Login Context (now local), Email and Password Input (local too)
+
 - 1.3.3
     - FIX: Frontend fix configs
+
 - 1.3.2
     - FIX: DB Test configs
+
 - 1.3.1
     - FIX: Local DB for Test
+
 - 1.3.0
     - ADD: README, CHANGELOG, LICENSE and .github folder
+
 - 1.2.3
-    - FIX: Show password error bugs 
+    - FIX: Show password error bugs
+
 - 1.2.2
     - FIX: Show email error bugs
+
 - 1.2.1
     - FIX: Enviroment API var URL
+
 - 1.2.0
     - ADD: Postgres and enviroment vars
     - FIX: Postgres server adding error
+
 - 1.1.1
     - FIX: Bugs in input types
     - ADD: ESlint ignore to RegEx
+
 - 1.1.0
     - ADD: Input Checker
     - FIX: Routes bugs
+
 - 1.0.0
     - ADD: Frontend and Backend